+++ CHANGELOG.md
@@ -12,6 +12,7 @@ miru.ink / Changelog
 - Update: Rename `size` property on ButtonGroup to `gap`.
 - Update: Allow component-type on `label` property on Button component + add `labelProps` property.
 - Update: New `reverse` property and stylings on ButtonGroup component.
+- Update: New `else` slot on and story content for BreakpointSupport component.
 - Fix: Include base stylings in dist (temporary solution).
 
 ## Version 0.0.8 - Experimental